Move UIError to output to avoid indirect dependency on UI in denoise
