+++ rebench/output.py
@@ -3,3 +3,22 @@ def output_as_str(string_like):
         return string_like.decode('utf-8')
     else:
         return string_like
+
+
+class UIError(Exception):
+
+    def __init__(self, message, exception):
+        super(UIError, self).__init__()
+        self._message = message
+        self._exception = exception
+
+    @property
+    def message(self):
+        return self._message
+
+    @property
+    def source_exception(self):
+        return self._exception
+
+    def __str__(self):
+        return self._message
